replaced old orchestrator.py with new
